Add multiple compilation flags
to make code even better

+++ src/fheroes2/ai/normal/ai_normal_kingdom.cpp
@@ -166,7 +166,7 @@ namespace AI
             }
         }
 
-        size_t heroLimit = world.w() / Maps::SMALL + 1;
+        int32_t heroLimit = world.w() / Maps::SMALL + 1;
         if ( _personality == EXPLORER )
             heroLimit++;
         if ( slowEarlyGame )
